Add commands table and update function list in README.md
Added a Commands section to the README file to provide a more clearer instructions on using the cat simulation program, each command now has a corresponding function in the table. Also updated the Functions section to reflect the command changes and current status of each function. These changes were made to increase user friendliness and to keep README file up to date.